Organize database interaction services
